Add files via upload
Warning! If you install this app with APK provided in the release section then you may have more trouble removing it on devices running Android P or earlier (easier on later version to remove and harder to configure as earlier versions require no configuration). Android Q and R do not allow hiding the app icon in the launcher and require overlay permission. Android Nougat, Oreo, and Pie is instant and will be difficult to remove as all Android versions before Q do not require a permission request and also do hide the app icon. I will not be held responsible for any damage.